commit_17.11.2021_20:10
Add a Builder pattern
Fix Repositories
Fix Swets
Delete DB

+++ NewYearPresent/NewYearPresent/Sweets/ISweetness.cs
@@ -4,7 +4,7 @@
     {
         string Name { get; set; }
         decimal Price { get; set; }
-        ushort Weight { get; set; }
+        float Weight { get; set; }
         float AmountOfSugar { get; set; }
     }
 }
